+++ stdlib/lib.core.d.ts
@@ -2639,6 +2639,7 @@ interface NetSocket {
     end(callback?: () => void, ...ignored: any[]): this;
     end(data: string | Buffer, callback?: () => void, ...ignored: any[]): this;
     destroy(callback?: () => void, ...ignored: any[]): this;
+    destroy(error: any, callback?: () => void, ...ignored: any[]): this;
     address(...ignored: any[]): SocketAddress | null;
     ref(...ignored: any[]): this;
     unref(...ignored: any[]): this;